Bug: ActiveRecord attributes have DB values
How to reproduce (there's a spec that demonstrates this)

1. Add an enumerized field that is an integer in the DB and a string in code. e.g. of this is user status in tests
2. Set the enumerized field using the string value
3. Reload the AR object
4. Note that the attributes hash will return the integer DB value

This is a regression from enumerize version 2.2.2. To be specific, this bug was introduced with this
PR https://github.com/brainspec/enumerize/pull/321, when a reload method was defined.

It seems that Rails calls Type#cast and this gem has not defined that method.
Defining cast resolves the issue. Type#deserialize by default calls Type#cast so removing that
definition should be ok.